[robust_estimation] inlier array => size_t to uint32_t #791

+++ src/openMVG/robust_estimation/rand_sampling_test.cpp
@@ -17,11 +17,11 @@ using namespace openMVG::robust;
 // Assert that each time exactly N random number are picked (no repetition)
 TEST(UniformSampleTestInARange, NoRepetions) {
 
-  std::vector<size_t> samples;
+  std::vector<uint32_t> samples;
   for (size_t total = 1; total < 500; total *= 2) { //Size of the data set
     for (size_t num_samples = 1; num_samples <= total; num_samples *= 2) { //Size of the consensus set
       UniformSample(num_samples, total, &samples);
-      const std::set<size_t> myset(samples.begin(), samples.end());
+      const std::set<uint32_t> myset(samples.begin(), samples.end());
       CHECK_EQUAL(num_samples, myset.size());
     }
   }
@@ -29,13 +29,13 @@ TEST(UniformSampleTestInARange, NoRepetions) {
 
 TEST(UniformSampleTestInAVectorOfUniqueElements, NoRepetions) {
 
-  std::vector<size_t> samples;
+  std::vector<uint32_t> samples;
   for (size_t total = 1; total < 500; total *= 2) { //Size of the data set
-    std::vector<size_t> vec_index(total);
+    std::vector<uint32_t> vec_index(total);
     std::iota(vec_index.begin(), vec_index.end(), 0);
     for (size_t num_samples = 1; num_samples <= total; num_samples *= 2) { //Size of the consensus set
       UniformSample(num_samples, &vec_index, &samples);
-      const std::set<size_t> myset(samples.begin(), samples.end());
+      const std::set<uint32_t> myset(samples.begin(), samples.end());
       CHECK_EQUAL(num_samples, myset.size());
     }
   }
